feat: move content to app
