+++ src/components/app/App.module.css
@@ -5,3 +5,10 @@
   height: 100%;
   background-color: #e5e5e5;
 }
+
+.main {
+  display: flex;
+  padding: 10px;
+  align-items: center;
+  flex-direction: column;
+}
