--- src/components/content/Content.module.css
@@ -1,6 +1,0 @@
-.main {
-  display: flex;
-  padding: 10px;
-  align-items: center;
-  flex-direction: column;
-}
